requirements.txt
livekit-agents
python-dotenv
aiohttp
librosa
tensorflow-cpu
numpy
soundfile
protobuf==5.28.3
opencv-python